fix: adapt to new yazi arrow command for pane navigation
Yazi's API has changed, requiring the use of the "arrow top" and "arrow bottom" commands
to jump to the top or bottom of the current pane. See the related change here:
https://github.com/sxyazi/yazi/pull/2294.

This commit updates the plugin to use the new commands, preventing the popup
from appearing when navigating panes in newer Yazi versions. :)

fix: wrong keyword